Release/v1.2

+++ tools/datasets/convert.py
@@ -94,20 +94,28 @@ def process_vidprom(root, info):
 
 def process_general_images(root, output):
     root = os.path.expanduser(root)
-    image_lists = get_filelist(root, IMG_EXTENSIONS)
-    df = pd.DataFrame(dict(path=image_lists))
-    if output is None:
-        output = "images.csv"
+    if not os.path.exists(root):
+        return
+    path_list = get_filelist(root, IMG_EXTENSIONS)
+    fname_list = [os.path.splitext(os.path.basename(x))[0] for x in path_list]
+    df = pd.DataFrame(dict(id=fname_list, path=path_list))
+
+    os.makedirs(os.path.dirname(output), exist_ok=True)
     df.to_csv(output, index=False)
     print(f"Saved {len(df)} samples to {output}.")
 
 
 def process_general_videos(root, output):
     root = os.path.expanduser(root)
-    video_lists = get_filelist(root, VID_EXTENSIONS)
-    df = pd.DataFrame(dict(path=video_lists))
-    if output is None:
-        output = "videos.csv"
+    if not os.path.exists(root):
+        return
+    path_list = get_filelist(root, VID_EXTENSIONS)
+    path_list = list(set(path_list))  # remove duplicates
+    fname_list = [os.path.splitext(os.path.basename(x))[0] for x in path_list]
+    relpath_list = [os.path.relpath(x, root) for x in path_list]
+    df = pd.DataFrame(dict(path=path_list, id=fname_list, relpath=relpath_list))
+
+    os.makedirs(os.path.dirname(output), exist_ok=True)
     df.to_csv(output, index=False)
     print(f"Saved {len(df)} samples to {output}.")
 
@@ -118,7 +126,7 @@ if __name__ == "__main__":
     parser.add_argument("root", type=str)
     parser.add_argument("--split", type=str, default="train")
     parser.add_argument("--info", type=str, default=None)
-    parser.add_argument("--output", type=str, default=None)
+    parser.add_argument("--output", type=str, default=None, required=True, help="Output path")
     args = parser.parse_args()
 
     if args.dataset == "imagenet":
